Fix compiler warnings

+++ cores/msp430/wiring_analog.c
@@ -139,7 +139,7 @@ void analogWrite(uint8_t pin, int val)
 		digitalWrite(pin, LOW); // set pin to LOW when duty cycle is 0
                                         // digitalWrite will take care of invalid pins
 	}
-	else if (val == analog_res)
+	else if ((uint16_t)val == analog_res)
 	{
 		digitalWrite(pin, HIGH); // set pin HIGH when duty cycle is 255
                                          // digitalWrite will take care of invalid pins
@@ -301,7 +301,7 @@ void analogWrite(uint8_t pin, int val)
  
                         case NOT_ON_TIMER:                      // not on a timer output pin
 			default:                                // or TxA0 pin
-				if (val <= (analog_res >> 1)) {
+				if ((uint16_t)val <= (analog_res >> 1)) {
 					digitalWrite(pin, LOW); // 
 				} else {
 					digitalWrite(pin, HIGH);
